requirements.txt
python-telegram-bot==21.5
apscheduler==3.10.4
pytz==2024.1
aiohttp==3.9.5
requests==2.31.0
qrcode==7.4.2
pillow==11.0.0
